requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
idna==3.7
load-dotenv==0.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.2
